requirements.txt
langchain
langchain-google-genai
python-telegram-bot
